# pkg: reduce the allocation of observe